Replace httpserver.c

+++ asgn4/httpserver.c
@@ -126,7 +126,7 @@ void *worker_threads(){
         fprintf(stderr, "thread  is in\n");
         uintptr_t conn = 0;
         queue_pop(task_queue,(void **)&conn);
-        int connfd = *(*(uintptr_t*)&conn);
+        uintptr_t connfd = *(*(uintptr_t*)&conn);
         fprintf(stderr, "conn: %d\n", connfd);
         fprintf(stderr, "queue pop is smooth\n");
         //handle_connection(connfd);
